Moved DPScopeApp() to its own module app.py
